showing isDirty, compile errors, class comments
Closes https://github.com/dalehenrich/Rowan/issues/160
Added bold to the italics displayed in a dirty project.
Method category indication no longer is italics and bold.

Closes https://github.com/dalehenrich/Rowan/issues/152
Method browser now shows compile errors properly

Closes https://github.com/dalehenrich/Rowan/issues/158
Saving class comments worked if you pressed ctrl-s but not if you used
the popup menu Save. Found that there was a high level Save context
menu in the system browser which pointed to savePackage. Replaced that
with a class comment context menu which saves the comment.

Removed tests for package policy in the browser. Needs further testing
after Dale gets checked in.