Fix BootJade tool hooks. Start JadeServer de-string
The development tools menu for Jadeite was not automatically setup so if you opened a Jadeite login shell from the Dolphin system Tools menu, you'd get the Jade login, not the Rowan. -- The former doesn't work.

Started implementing a better solution to the Jade string. Most work done in Rowan Services on the server.