Refactor update trigger/Method browser updates
Refactored trigger/updates for registering presenters and sending updates.
Was able to get rid of a bunch of similar methods in BrowserUpdate.

Method source updates properly color the pane after successful method
completion. More important, the update is setup for method source changes.

Removed selection of first project automatically in browser.

Moved the update of the Projects tab so that it is called regardless of
whether there are selections or not.

Moved the registering of presenters to a better spot (at least I think)