Fix dependency leak in update architecture
The service update architecture was setup inefficiently such that the BrowserUpdate was retaining dependencies for presenters long dead. I had setup the BrowserUpdate as the trigger from & target. Since the BrowserUpdate needs to live beyond the life of a browser, this created a problem. Moving the target of the dependency to the presenter fixed the problem even if there is a touch more overhead.

(That overhead may be able to be cleaned up with more analysis)

Also minor cleanup to avoid a round-trip when getting the default class template for a package.