Client updates more selective
The server now can tell the client what update (selector) to run when it sends back an update. The client will respect that.

Also removed all the recently added update methods in Presenter in lieu of a single method - This is simpler.

See RowanService>>updateType for a first attempt at using it - Two separate browsers registered for method removals and if you remove a method in the method list, it will remove it in the main browser as well.