Fix #130. To login without code modification privilege, you will need to install JadeServer: * Using Jade, log in to a stone with the same version (e.g., 3.2.15) where you have code modification privilege (such as a different user on the same stone); * From the Jade Transcript, select the Jade menu and the 'File out JadeServer' menu command; * Save the text file to a convenient location; * Using Topaz, log in to the stone as SystemUser (or edit the file to select another SymbolDictionary, such as Published, and login as another user, e.g., DataCurator); * Input the file and commit; and * Using Jade, log in and things should work properly.
The requirement to file out from a logged-in session allows us to use the server code that exports in a Topaz format (this isn't how we internally create JadeServer on login if it isn't present).

If the installed version of JadeServer does not match the client version of Jade (see the 'About Jade' menu) you will be asked whether to use the installed JadeServer. Do so at your own risk!